+++ routes/api/contacts.js
@@ -1,5 +1,7 @@
 const express = require('express');
-const ctrl = require("../../controllers/contacts")
+const ctrl = require("../../controllers/contacts");
+const { validateBody } = require("../../middlewares");
+const schema = require("../../schemas/contacts")
 
 const router = express.Router();
 
@@ -7,7 +9,7 @@ router.get('/', ctrl.getAll);
 
 router.get('/:id', ctrl.getById);
 
-router.post('/', ctrl.addCont)
+router.post('/', validateBody(schema.schema),  ctrl.addCont)
 
 router.delete('/:id', ctrl.deleteContact)
